backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
black==24.4.2
boto3==1.34.131
botocore==1.34.131
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email-validator==2.1.1
fastapi==0.110.0
filelock==3.14.0
flake8==7.0.0
frozenlist==1.4.1
fsspec==2024.3.1
google-api-core==2.17.1
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
librt==0.7.8
litellm==1.35.32
markdown-it-py==3.0.0
MarkupSafe==2.1.5
motor==3.4.0
multidict==6.0.5
mypy==1.10.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.30.1
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
passlib==1.7.4
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.1
pydantic-core==2.18.2
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.7.2
pyparsing==3.1.2
pytest==8.2.1
pytest-asyncio==0.23.6
prometheus-fastapi-instrumentator==7.0.0
redis==5.0.4
locust==2.27.0
mongomock-motor==0.0.33
python-dateutil==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
razorpay==1.4.2
regex==2024.5.15
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
scikit-learn==1.4.2
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
stripe==8.8.0
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
zipp==3.18.1
reportlab==4.4.10
